+++ Modules/Fonts.lua
@@ -5,6 +5,7 @@
 FONT_PATH_SETTINGS = {
     { "hpRatioFontKey", "hpRatioFontPath" },
     { "nameFontKey", "nameFontPath" },
+    { "castbarSpellNameFontKey", "castbarSpellNameFontPath" },
     { "levelOverlayFontKey", "levelOverlayFontPath" },
     { "chatFontKey", "chatFontPath" },
 }